:sparkles: Fix many clang-tidy warnings

+++ plugins/csp-lod-bodies/src/TileVisitor.hpp
@@ -11,8 +11,6 @@
 #include "TileId.hpp"
 #include "TileQuadTree.hpp"
 
-#include <boost/assert.hpp>
-
 namespace csp::lodbodies {
 
 class TileNode;
@@ -239,7 +237,7 @@ void TileVisitor<DerivedT>::visitRoot(TileNode* rootDEM, TileNode* rootIMG, Tile
   state.mTileId      = tileId;
 
   if (self().preVisitRoot(tileId)) {
-    for (int i = 0; i < 4; ++i) {
+    for (size_t i = 0; i < 4; ++i) {
       if (!state.mChildren[i]) {
         continue;
       }
@@ -277,7 +275,7 @@ void TileVisitor<DerivedT>::visitLevel(TileNode* nodeDEM, TileNode* nodeIMG, Til
   state.mTileId      = tileId;
 
   if (self().preVisit(tileId)) {
-    for (int i = 0; i < 4; ++i) {
+    for (size_t i = 0; i < 4; ++i) {
       if (!state.mChildren[i]) {
         continue;
       }
